% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamStats_Delineation_single.R
\name{streamStats_Delineation_single}
\alias{streamStats_Delineation_single}
\title{StreamStats Autodelineate a Single Watershed}
\usage{
streamStats_Delineation_single(state, longitude, latitude, UID)
}
\arguments{
\item{state}{USGS StreamStats state info e.g. 'VA'}

\item{longitude}{Decimal Degree Longitude value, numeric}

\item{latitude}{Decimal Degree Latitude value, numeric}

\item{UID}{Unique station identifier to append to each feature}
}
\value{
A list object with two elements, a polygon and a point
}
\description{
This function establishes a standardized, automated method for scraping
the USGS StreamStats API to delineate pour point watersheds and organizing them into
a single shapefile. This method replaces the arduous and inconsistent task of manually
delineating watersheds using ESRI mapping products.
This script was built with example code from Ryan King
https://ryan-hill.github.io/sfs-r-gis-2018/modules/rasters/extract-raster-data-r/ exercise 3
and modified by Emma Jones to fit DEQ's needs.
}
\examples{
streamStats_Delineation_single(state = "VA", longitude = -80.063946, latitude = 37.812840, UID = "Example Site")
}
